PR comments. New actions, clean expands, add URI lib
