+++ gatsby-node.js
@@ -91,7 +91,7 @@ const SSR_getSpeakers = async (baseUrl, summitId, accessToken, filter = null, pa
 
 const SSR_getSummit = async (baseUrl, summitId) => {
     const params = {
-        expand: 'event_types,tracks,track_groups,presentation_levels,locations.rooms,locations.floors,order_extra_questions.values,order_extra_questions.sub_question_rules,order_extra_questions.sub_question_rules.answer_values,schedule_settings,schedule_settings.filters,schedule_settings.pre_filters',
+        expand: 'event_types,tracks,track_groups,presentation_levels,locations.rooms,locations.floors,order_extra_questions.values,schedule_settings,schedule_settings.filters,schedule_settings.pre_filters',
         t: Date.now()
     };
