+++ package.json
@@ -61,7 +61,7 @@
     "netlify-cms-lib-widgets": "^1.3.0",
     "node-sass": "^4.14.1",
     "node-sass-utils": "^1.1.3",
-    "openstack-uicore-foundation": "^3.0.60",
+    "openstack-uicore-foundation": "^3.0.61",
     "prop-types": "^15.6.0",
     "react": "^16.8.4",
     "react-avatar-editor": "11.0.9",
